feat(Spreadsheet manager): refactor spreadsheet writer
