+++ lib/lita-lunch-reminder.rb
@@ -11,7 +11,7 @@ require "lita/handlers/api/api_controller"
 require "lita/handlers/api/karma"
 require "lita/handlers/api/lunch"
 require "lita/handlers/api/market"
-require "lita/services/spreadsheet_writer"
+require "lita/services/spreadsheet_manager"
 require "lita/services/lunch_assigner"
 require "lita/services/weighted_picker"
 require "lita/services/karmanager"
